feature: add font resources in the application theme
The selected font for the application is Poppins. Import the bold, regular and thin types of font, and then create the Font Family for allowing Poppins to be used inside the application.
The Font Family has been included in the application Typography, which is part of the main theme.

Integrates #1