feature: build the list closed layout
Using Jetpack Compose, provide a unique implementationn on the list, showing the closed or the opened item based on a state variable.
Start with all the items closed and then open them upon clicking on it.
Delete all the previews related to the display of a single list item, opened or closed.

Integrates #5