feature: provide animation to the list item opening
Using Jetpack Compose, provide the implementation of the animation in the UI as soon as a specific item is clicked.
For it, wrap the implementation of a closed or opened item in a Box, and provide the animation inside that Box. Bring the card back to the parent function in order to let the Box animate.

Integrates #7