feature: generate the app bar
Place the appbar in the Scaffold composable and generate the app bar with the Image and the text centered.
Update the necessary string resources in the strings.xml file.
Insert a new parameter in the PlacesList composable function due to the fact that is called inside the Scaffold composable.

Integrates #11